# improve `LIKE` regex performance up to 12x