# Which issue does this PR close?

While working on #6118, I realised we can significantly improve the `Regex` branch of `Predicate` for `LIKE` and `ILIKE` queries.

# Rationale for this change

Significant improve in performance of `LIKE` queries:

<details>
<summary>Running this script</summary>

```bash
#!/usr/bin/env bash
set -ex
if [ "$#" -ne 2 ]; then
  echo "Usage: $0 branch-name bench-filter"
  exit 1
fi
BRANCH_NAME=$1
BENCH_FILTER=$2
git checkout master
rm -rf target/criterion
cargo bench -p arrow --bench comparison_kernels -F test_utils -- --save-baseline master $BENCH_FILTER
git checkout $BRANCH_NAME
cargo bench -p arrow --bench comparison_kernels -F test_utils -- --save-baseline $BRANCH_NAME $BENCH_FILTER
critcmp master $BRANCH_NAME
```

</details>

With `./bench.sh improve-like-regex like` on `c2-standard-8` GCP instance.

### with hackstack length 4 (default)

```
+ critcmp master improve-like-regex
group                                              improve-like-regex      master
-----                                              ------------------      ------
ilike_utf8 scalar complex                          1.17    358.8±5.76µs    1.00    307.0±1.44µs
ilike_utf8 scalar contains                         1.00  1161.5±17.13µs    1.31   1516.2±4.97µs
ilike_utf8 scalar ends with                        1.05    246.9±2.88µs    1.00    234.6±0.59µs
ilike_utf8 scalar equals                           1.16    249.1±3.29µs    1.00    214.3±0.51µs
ilike_utf8 scalar starts with                      1.11    279.7±3.69µs    1.00    251.4±0.45µs
ilike_utf8_scalar_dyn dictionary[10] string[4])    1.00     49.1±0.67µs    1.79     88.0±0.20µs
like_utf8 scalar complex                           1.08    311.4±1.78µs    1.00    288.5±0.81µs
like_utf8 scalar contains                          1.01    376.5±1.01µs    1.00    374.0±0.66µs
like_utf8 scalar ends with                         1.15    268.8±0.52µs    1.00    234.7±0.90µs
like_utf8 scalar equals                            1.25    271.4±0.61µs    1.00    217.4±0.45µs
like_utf8 scalar starts with                       1.00    234.5±0.79µs    1.08    253.9±1.13µs
like_utf8_scalar_dyn dictionary[10] string[4])     1.00     49.2±0.24µs    1.79     88.2±0.42µs
like_utf8view scalar complex                       1.00    169.4±1.01ms    3.13    529.6±3.51ms
like_utf8view scalar contains                      1.00    384.0±0.43ms    1.04    400.1±0.58ms
like_utf8view scalar ends with                     1.00     53.6±0.19ms    1.10     59.0±0.20ms
like_utf8view scalar equals                        1.03     30.3±0.08ms    1.00     29.4±0.10ms
like_utf8view scalar starts with                   1.00     55.0±0.24ms    1.01     55.5±0.20ms
nilike_utf8 scalar complex                         1.16    358.4±5.66µs    1.00    307.9±1.05µs
nilike_utf8 scalar contains                        1.00  1160.2±15.34µs    1.31   1517.2±5.90µs
nilike_utf8 scalar ends with                       1.05    246.6±1.00µs    1.00    234.5±0.51µs
nilike_utf8 scalar equals                          1.16    249.1±3.31µs    1.00    214.2±0.42µs
nilike_utf8 scalar starts with                     1.11    279.8±3.45µs    1.00    251.5±0.73µs
nlike_utf8 scalar complex                          1.09    313.4±4.82µs    1.00    287.7±0.75µs
nlike_utf8 scalar contains                         1.01    376.0±1.23µs    1.00    373.7±0.58µs
nlike_utf8 scalar ends with                        1.15    268.8±1.15µs    1.00    234.5±0.39µs
nlike_utf8 scalar equals                           1.25    271.6±3.42µs    1.00    217.3±0.48µs
nlike_utf8 scalar starts with                      1.00    234.9±2.93µs    1.08    253.7±0.48µs
```

### with haystack length 0..400

As described in https://github.com/apache/arrow-rs/pull/6128#issuecomment-2254513573

```
+ critcmp master improve-like-regex
group                                              improve-like-regex      master
-----                                              ------------------      ------
ilike_utf8 scalar complex                          1.00      2.4±0.06ms    9.45     23.0±0.08ms
ilike_utf8 scalar contains                         1.00      3.8±0.04ms    6.02     22.9±0.06ms
ilike_utf8 scalar ends with                        1.00   970.7±43.73µs    1.21  1171.2±51.28µs
ilike_utf8 scalar equals                           1.00   626.4±24.91µs    1.21   760.1±36.51µs
ilike_utf8 scalar starts with                      1.00  1025.3±22.76µs    1.12  1150.5±83.44µs
ilike_utf8_scalar_dyn dictionary[10] string[4])    1.79     88.2±0.17µs    1.00     49.3±0.14µs
like_utf8 scalar complex                           1.00  1707.9±24.96µs    12.76    21.8±0.09ms
like_utf8 scalar contains                          1.00      5.3±0.03ms    1.01      5.4±0.02ms
like_utf8 scalar ends with                         1.02   480.0±15.99µs    1.00   472.4±11.46µs
like_utf8 scalar equals                            1.00     90.7±0.24µs    1.39    126.4±0.18µs
like_utf8 scalar starts with                       1.00    496.5±9.34µs    1.03   509.9±14.96µs
like_utf8_scalar_dyn dictionary[10] string[4])     1.79     87.9±0.14µs    1.00     49.1±0.15µs
like_utf8view scalar complex                       1.00    178.5±0.35ms    3.00    535.2±3.98ms
like_utf8view scalar contains                      1.04    400.0±0.42ms    1.00    385.9±0.75ms
like_utf8view scalar ends with                     1.04     56.8±0.19ms    1.00     54.8±0.23ms
like_utf8view scalar equals                        1.00     29.8±0.09ms    1.02     30.5±0.13ms
like_utf8view scalar starts with                   1.00     55.3±0.21ms    1.02     56.1±0.27ms
nilike_utf8 scalar complex                         1.00      2.4±0.07ms    9.58     22.9±0.09ms
nilike_utf8 scalar contains                        1.00      3.8±0.04ms    6.02     22.8±0.10ms
nilike_utf8 scalar ends with                       1.00   980.9±20.47µs    1.05  1034.4±26.55µs
nilike_utf8 scalar equals                          1.00   651.9±41.19µs    1.07   698.9±24.53µs
nilike_utf8 scalar starts with                     1.00  1034.2±18.20µs    1.12  1162.5±52.16µs
nlike_utf8 scalar complex                          1.00  1696.5±43.69µs    12.86    21.8±0.09ms
nlike_utf8 scalar contains                         1.00      5.3±0.01ms    1.02      5.4±0.03ms
nlike_utf8 scalar ends with                        1.00   468.4±16.73µs    1.04   485.3±18.10µs
nlike_utf8 scalar equals                           1.00     90.7±0.15µs    1.39    126.5±0.23µs
nlike_utf8 scalar starts with                      1.00    498.8±9.86µs    1.04   521.1±12.16µs
```

# What changes are included in this PR?

Changed `regex_like` so that leading and trailing `%` are simply removed instead of being replaced with `^.*` and `.*$` respectively, so `LIKE '%foo%'` becomes `Regex('foo')` instead of `Regex('^.*foo.*$')`.

# Are there any user-facing changes?

Since we had `.dot_matches_new_line(true)` I believe this is the exact same behaviour as we had before, just much faster.