@samuelcolvin 2024-07-29
> Nice find, surprised it is faster (would have thought it would be optimized away).

Ye I was surprised too. It makes sense that if you don't have `.dot_matches_new_line(true)`, you need to iterate through all of `.*` to check validity.