@samuelcolvin 2024-07-29
> I know one of the ClickBench queries has a giant LIKE that dominates performance on several -- e.g. https://github.com/apache/datafusion/blob/ea8c287855ae799573a56697db63e7b16edf3a37/benchmarks/queries/clickbench/queries.sql#L21-L24

Since that will already use `Predicate::Contain`, that case should be improved by #6128, not this.

This case will help with more complex `LIKE` queries, and `ILIKE` contains queries.